+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 3.0.0 - 2019-07-31
+
+* Required PHP Version raised to 7.1.3
+* Factory replaced by Builder
+* Cache mechanism (PSR-16 Compatible) (NullCache by default)
+* Logger mechanism (PSR-3 Compatible) (NullLogger by default)
+* Docker image for development purposes
+* Minor changes
+
 ## 2.0.0 - 2018-07-20
 
 * Factory extracted from main validator class (single responsibility principle)